+++ web/pages/feeds/export.svelte
@@ -1,18 +1,12 @@
 <script lang="ts">
-  import {
-    commonMessages as common,
-    exportMessages,
-    getCategories,
-    getFeeds,
-    exportMessages as t
-  } from '@slowreader/core'
+  import { exportMessages as t } from '@slowreader/core'
 
   import { getURL } from '../../stores/router.js'
   import CardLink from '../../ui/card-link.svelte'
   import CardLinks from '../../ui/card-links.svelte'
   import TwoStepsPage from '../../ui/two-steps-page.svelte'
-  import ExportInternal from './exportInternal.svelte'
-  import ExportOPML from './exportOPML.svelte'
+  import ExportInternal from './internal.svelte'
+  import ExportOPML from './opml.svelte'
 
   export let formatId: string | undefined
   const formats = [
